Escape user input in Settings fields.
A malicious Administrator could enter HTML into the text fields to break out of them and execute arbitrary JavaScript. For example, "><img src=x onerror=prompt(1)>

Because this requires an Administrator account in order to exploit, it is a minor vulnerability and unlikely to be a useful attack vector.
props trizaeron
xref: https://hackerone.com/reports/9375


git-svn-id: https://plugins.svn.wordpress.org/basic-google-maps-placemarks/branches/1.10@901237 b8457f37-d9ea-0310-8a92-e5e31aec5664